refactor: Private, Public Router 제거 및 LayoutProvider로 수정
