+++ package.json
@@ -15,33 +15,33 @@
     "build-storybook": "storybook build"
   },
   "dependencies": {
-    "@tanstack/react-query": "^4.35.0",
-    "axios": "^1.5.0",
-    "jotai": "^2.4.2",
+    "@tanstack/react-query": "^4.35.7",
+    "axios": "^1.5.1",
+    "jotai": "^2.4.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.15.0",
     "styled-components": "^6.0.7",
-    "swiper": "^10.3.0"
+    "swiper": "^10.3.1"
   },
   "devDependencies": {
-    "@storybook/addon-essentials": "^7.4.1",
-    "@storybook/addon-interactions": "^7.4.1",
-    "@storybook/addon-links": "^7.4.1",
+    "@storybook/addon-essentials": "^7.4.5",
+    "@storybook/addon-interactions": "^7.4.5",
+    "@storybook/addon-links": "^7.4.5",
     "@storybook/addon-onboarding": "^1.0.8",
-    "@storybook/addon-viewport": "^7.4.1",
-    "@storybook/blocks": "^7.4.1",
-    "@storybook/react": "^7.4.1",
-    "@storybook/react-vite": "^7.4.1",
-    "@storybook/testing-library": "^0.2.0",
+    "@storybook/addon-viewport": "^7.4.5",
+    "@storybook/blocks": "^7.4.5",
+    "@storybook/react": "^7.4.5",
+    "@storybook/react-vite": "^7.4.5",
+    "@storybook/testing-library": "^0.2.2",
     "@testing-library/jest-dom": "^6.1.3",
     "@testing-library/react": "^14.0.0",
     "@types/jest": "^29.5.4",
-    "@types/react": "^18.2.15",
-    "@types/react-dom": "^18.2.7",
-    "@typescript-eslint/eslint-plugin": "^6.6.0",
-    "@typescript-eslint/parser": "^6.6.0",
-    "@vitejs/plugin-react": "^4.0.3",
+    "@types/react": "^18.2.24",
+    "@types/react-dom": "^18.2.8",
+    "@typescript-eslint/eslint-plugin": "^6.7.4",
+    "@typescript-eslint/parser": "^6.7.4",
+    "@vitejs/plugin-react": "^4.1.0",
     "eslint": "^8.49.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.1.0",
@@ -60,15 +60,15 @@
     "jest-environment-jsdom": "^29.6.4",
     "jest-svg-transformer": "^1.0.0",
     "lint-staged": "^14.0.1",
-    "msw": "^1.3.0",
+    "msw": "^1.3.2",
     "prettier": "^3.0.3",
-    "storybook": "^7.4.1",
+    "storybook": "^7.4.5",
     "storybook-addon-breakpoints": "^1.0.0",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
     "typescript": "^5.0.2",
-    "vite": "^4.4.5",
-    "vite-plugin-svgr": "^3.2.0"
+    "vite": "^4.4.10",
+    "vite-plugin-svgr": "^4.1.0"
   },
   "msw": {
     "workerDirectory": "public"
