+++ src/App.tsx
@@ -1,7 +1,7 @@
+import LayoutProvider from './provider/LayoutProvider';
 import { Outlet } from 'react-router-dom';
 import { ThemeProvider } from 'styled-components';
 import { setScreenSize } from './utils/styles';
-import { styled } from 'styled-components';
 import { themeDefault } from './styles/theme';
 import { useEffect } from 'react';
 
@@ -16,25 +16,11 @@ function App() {
 
   return (
     <ThemeProvider theme={themeDefault}>
-      <Layout>
-        <Container>
-          <Outlet />
-        </Container>
-      </Layout>
+      <LayoutProvider>
+        <Outlet />
+      </LayoutProvider>
     </ThemeProvider>
   );
 }
 
 export default App;
-
-const Layout = styled.section`
-  display: flex;
-  flex-direction: column;
-  background-color: rgb(245, 245, 245);
-`;
-
-const Container = styled.section`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
