--- src/router/PublicProvider.tsx
@@ -1,13 +1,0 @@
-type Props = {
-  children: React.ReactNode;
-}
-
-const PublicProvider = ({ children }: Props) => {
-  return (
-    <>
-      {children}
-    </>
-  );
-}
-
-export default PublicProvider;
