+++ src/router/index.tsx
@@ -2,8 +2,6 @@ import AlbumDetailPage from '../pages/AlbumDetail';
 import App from '../App';
 import HomePage from '../pages/Home';
 import LoginPage from '../pages/Login';
-import PrivateProvider from './PrivateProvider';
-import PublicProvider from './PublicProvider';
 import { ROUTES_PATH } from '../constants/routes';
 import Redirection from './Redirection';
 import TestPage from '../pages/Test';
@@ -16,35 +14,19 @@ const router = createBrowserRouter([
     children: [
       {
         index: true,
-        element: (
-          <PrivateProvider>
-            <HomePage />
-          </PrivateProvider>
-        ),
+        element: <HomePage />,
       },
       {
         path: ROUTES_PATH.albumDetail,
-        element: (
-          <PrivateProvider>
-            <AlbumDetailPage />
-          </PrivateProvider>
-        ),
+        element: <AlbumDetailPage />,
       },
       {
         path: ROUTES_PATH.login,
-        element: (
-          <PublicProvider>
-            <LoginPage />
-          </PublicProvider>
-        ),
+        element: <LoginPage />,
       },
       {
         path: '/test',
-        element: (
-          <PrivateProvider>
-            <TestPage />
-          </PrivateProvider>
-        ),
+        element: <TestPage />,
       },
     ],
   },
